+++ test/crypt-utils.test.ts
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-// @ts-ignore
-import * as HDKey from 'hdkey'
 import * as chai from 'chai'
 import * as chaiAsPromised from 'chai-as-promised'
 import * as sinon from 'sinon'
 import { CryptUtil, LocalCryptUtils } from '../src'
+import { ethers } from 'ethers'
 
 const assert = chai.assert
 
@@ -61,23 +60,127 @@ describe('cryptutils class', () => {
 
   it('should import a masterprivatekey', () => {
     // Arrange
-    const stub = sinon.stub(HDKey, 'fromExtendedKey')
-    const privKey = 'fromExtendedKey'
+    const stub = sinon.stub(ethers.utils.HDNode, 'fromExtendedKey')
+    const privKey = 'xprv9s21ZrQH143K4Hahxy3chUqrrHbCynU5CcnRg9xijCvCG4f3AJb1PgiaXpjik6pDnT1qRmf3V3rzn26UNMWDjfEpUKL4ouy6t5ZVa4GAJVG'
     // Act
     sut.importMasterPrivateKey(privKey)
     // Assert
     assert.isTrue(stub.calledOnceWithExactly(privKey))
   })
 
-  it('should throw an error if hdkey cannot create a masterprivatekey', () => {
+  it('should throw an error if hdnode cannot create a masterprivatekey', () => {
     // Arrange
-    sinon.stub(HDKey, 'fromMasterSeed').returns(false)
+    sinon.stub(ethers.utils.HDNode, 'fromSeed')
     const helper = function () {
       sut.createMasterPrivateKey()
     }
 
     // Act / Assert
     assert.throws(helper, 'Could not create master private key')
+  })
+
+  it('should throw if function exportMasterPrivateKey called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.exportMasterPrivateKey()
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function derivePrivateKey called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.derivePrivateKey(0, 0)
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function derivePublicKey called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.derivePublicKey(0, 0)
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function deriveAddress called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.deriveAddress(0, 0)
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function derivePublicExtendedKey called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.derivePublicExtendedKey(0, 0)
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function derivePublicExtendedKeyFromPath called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.derivePublicExtendedKeyFromPath('')
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function derivePrivateKeyFromPath called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.derivePrivateKeyFromPath('')
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
+  })
+
+  it('should throw if function signPayload called without instantiating a masterprivatekey', () => {
+    // Arrange
+    const sut = new LocalCryptUtils()
+
+    // Act
+    const helper = () => {
+      sut.signPayload(0, 0, '')
+    }
+
+    // Assert
+    assert.throws(helper, 'No MasterPrivateKey instantiated')
   })
 
   it('should should return a corresponding private extended key', () => {
@@ -109,7 +212,25 @@ describe('cryptutils class', () => {
     assert.equal(pubExtendedKey.length, 111)
   })
 
-  it('should calculate the address out of a public key', () => {
+  it('should calculate the address out of a public key prefixed with 0x04', () => {
+    // Arrange
+    const pubkey = '0x045c1d9376dd92af86696de24806477a40c21291831840a220da1eac511c758c28553456e13ea0057641aa2dc4e66cfffbd49ae3a316f933f6613f87bf7e8fdf77'
+    const matchingAddress = '0xdA2B12cED8B2fc19c5abEF68aC99F55616BC98eB'
+
+    // Act & Assert
+    assert.equal(sut.getAddressFromPubKey(pubkey), matchingAddress)
+  })
+
+  it('should calculate the address out of a public key prefixed with 04', () => {
+    // Arrange
+    const pubkey = '045c1d9376dd92af86696de24806477a40c21291831840a220da1eac511c758c28553456e13ea0057641aa2dc4e66cfffbd49ae3a316f933f6613f87bf7e8fdf77'
+    const matchingAddress = '0xdA2B12cED8B2fc19c5abEF68aC99F55616BC98eB'
+
+    // Act & Assert
+    assert.equal(sut.getAddressFromPubKey(pubkey), matchingAddress)
+  })
+
+  it('should calculate the address out of a public key not prefixed at all', () => {
     // Arrange
     const pubkey = '5c1d9376dd92af86696de24806477a40c21291831840a220da1eac511c758c28553456e13ea0057641aa2dc4e66cfffbd49ae3a316f933f6613f87bf7e8fdf77'
     const matchingAddress = '0xdA2B12cED8B2fc19c5abEF68aC99F55616BC98eB'